+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # ChangeLog
 
 * fix recipetype project doesn't show up the correct icon
+* fix canceling recipe event doesn't work
+* bump up the ldlib for QoL
+* add the day light condition